+++ src/AngleSharp.Css/Dom/ICssProperty.cs
@@ -57,5 +57,12 @@ namespace AngleSharp.Css.Dom
         /// Gets if the property is a shorthand.
         /// </summary>
         Boolean IsShorthand { get; }
+
+        /// <summary>
+        /// Creates a computed version of the property.
+        /// </summary>
+        /// <param name="device">The device to compute for.</param>
+        /// <returns>The computed version of the property if uncomputed, otherwise the same.</returns>
+        ICssProperty Compute(IRenderDevice device);
     }
 }