+++ src/AngleSharp.Css/Values/Primitives/Angle.cs
@@ -6,7 +6,7 @@ namespace AngleSharp.Css.Values
     /// Represents an angle object.
     /// https://developer.mozilla.org/en-US/docs/Web/CSS/angle
     /// </summary>
-    public struct Angle : IEquatable<Angle>, IComparable<Angle>, ICssPrimitiveValue
+    public readonly struct Angle : IEquatable<Angle>, IComparable<Angle>, ICssPrimitiveValue
     {
         #region Basic angles